Fixed all students data table - changed column ID to match column name - renamed file to match page title - added label to delete row action - replaced mass action button with action menu - removed unnecessary prop drilling
